Reformat Nowsecure output and introduce the GH action (#3)
This PR attempts to make the action a bit more generic in order to be used for the rest of the teams in the account as well. Also, it introduces a Dockerfile and the relevant action.yaml file in order to be utilised in the pipelines of each GH repo.