feat: Populating the NowSecure pdf report (#135)
This PR adds the ability to generate the pdf report per platform for the latest assessment retrieved from the action. Currently, the report(s) will reside in the filesystem on which the action will run on. In an upcoming PR, the ability to export that report to a chained action will be added.

Contributes to: camelotls/actions-nowsecure#130

Signed-off-by: Stelios Gkiokas <stelios.giokas@camelotls.com>